+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## feat-poll_records (wip)
+## feat-refactor_poll_providers (wip)
 
    ### Added
      - (poll) Added File provider to read YAML/JSON from filesystem
@@ -6,7 +6,7 @@
      - (poll) Added Remote provider to read YAML/JSON from a HTTP/HTTPS source with basic authentication supported
 
    ### Changed
-     - Created common functions for poll providers for easier implementation of future pollers
+     - Created pollCommonfunctions for poll providers (http, records management, options, processing of parsed,received data, filter logic for easier implementation of future pollers)
      -
 
 ## 1.0.0 2025-05-23 <code at nfrastack dot com>